fix: 웨이팅 알림톡 페이지 수정
